// client/src/app/api/models/grand-prix-championship-context-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { SessionDefinition } from '../models/session-definition';
export interface GrandPrixChampionshipContextDto {
  availableSessions?: Array<SessionDefinition> | null;
  categories?: Array<string> | null;
  consChampId?: string | null;
  driversChampId?: string | null;
  pointSystem?: string | null;
}