// client/src/app/api/models/batch-result-create-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { SingleResultDto } from './single-result-dto';
export interface BatchResultCreateDto {
  consChampId?: string;
  driversChampId?: string;
  grandPrixId?: string;
  results?: Array<SingleResultDto> | null;
  session?: string | null;
}