// client/src/app/api/models/wec-batch-result-create-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { WecSingleResultDto } from '../models/wec-single-result-dto';
export interface WecBatchResultCreateDto {
  consChampId?: string;
  driversChampId?: string | null;
  grandPrixId?: string;
  results?: Array<WecSingleResultDto> | null;
  session?: string | null;
}